Ghost AI Changes
This is the minor change for the project. I have made it so two of the ghosts have more waypoints which make their patrols more dynamic. The final ghost has a set range for a random patrol and will walk around the right side of the house to random spots. I also made the ghost in the kitchen twice as fast and changed his texture to make him stand out as a faster enemy type.